Finish Input Handling section of blogpost
- Cross section selected on click
- Tile follows mouse for dragging
- Input adapter wired in
- Stare of Match3GameState started.

The match3 game state might get some heavy refactoring later.
Tempting to make an abstract base class to define empty interface
methods and then inherit instead of using the enums, that would
make it possible to have the idle state be a singleton, the in
progress state to track its state separately, and the end state
to do whatever, but eh, for now it's fine as is.