Add the start of match processing
- BoardAcceptingMoves state to handle all the state we wrote before about moves
- BoardState interface for all the states we can swap between
- ClearGridSpace command
- helper in GameGrid to get clear commands for matches
- Fix for GridSpace to avoid nulls being true, Object.equals does that and we don't want that
- Match3State swaps between processing and accepting move states, also broadcasts matches made
- Pub Sub interfaces for matches
- ProcessingMatches start that clears tiles out and then gets stuck in that state for now.

This all took way too much brain power to put together. I don't know why my head was
moving through molasses today but it was really frustrating. I still need to deal with:

1. Shift tiles for gravity
2. Replace tiles with new tiles from the token generator
3. recheck for matches and then exit state as needed

And once that's all done we'll need to add in the score panel. And at that point
we'll need to swap from crappy tiles over to using the asset manager and real
graphics. Then after that we'll add in sound. Before the swap we'll add in a
loading screen and title screen as well I suppose and then at that point I think
I'll be even with my initial prototype as far as features go.

Although I need to refactor the command structure of the tile graphics I think.
the handle commmand thing just seems wrong.