Scoring graphic and calculator
- I should probably make the scoring calculator return the current score and
multiplier because that would make it slightly easier to swap out if we ever
wanted to have a hard mode/easy mode difference in scoring.
- For now the ScoreGraphic listens to TileType matches but I'm probably going
to change that over to a different type so that I can avoid the usage of
BoardGraphic inside of the score graphic because I really don't like that.